added store area,  product editing, image upload, minor admin tweaks

+++ Program.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using semenarna_id2.Data;
+using semenarna_id2.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -27,24 +28,44 @@ namespace semenarna_id2 {
 
                 var adminRole = new IdentityRole("Admin");
                 var userRole = new IdentityRole("User");
+                roleMgr.CreateAsync(userRole).GetAwaiter().GetResult();
 
                 if (!ctx.Roles.Any()) {
                     //create role
                     roleMgr.CreateAsync(adminRole).GetAwaiter().GetResult();
                 }
-                /*if(ctx.Users.Any(u => u.UserName == "admin")) {*/
-                /*var usr = userMgr.FindByNameAsync("admin").GetAwaiter().GetResult();
-                var res = userMgr.DeleteAsync(usr).GetAwaiter().GetResult();*/
+                if (!ctx.Users.Any(u => u.UserName == "admin"))
+                {
+                    /*var usr = userMgr.FindByNameAsync("admin").GetAwaiter().GetResult();
+                    var res = userMgr.DeleteAsync(usr).GetAwaiter().GetResult();*/
 
-                //create admin
-                /*var adminUser = new IdentityUser {
-                    UserName = "admin",
-                    Email = "admin@paralax.mk"
-                };
-                var result = userMgr.CreateAsync(adminUser, "radant098").GetAwaiter().GetResult();
-                userMgr.AddToRoleAsync(adminUser, adminRole.Name).GetAwaiter().GetResult();*/
+                    //create admin
+                    var adminUser = new IdentityUser
+                    {
+                        UserName = "admin",
+                        Email = "admin@paralax.mk"
+                    };
+                    var result = userMgr.CreateAsync(adminUser, "radant098").GetAwaiter().GetResult();
+                    userMgr.AddToRoleAsync(adminUser, adminRole.Name).GetAwaiter().GetResult();
 
-                /*}*/
+                }
+                if (!ctx.TestProduct.Any()) {
+                    var seedList = new List<TestProductModel>();
+
+                    for(int i = 0; i < 6; i++) {
+                        var product = new TestProductModel {
+                            Name = "Test Product " + i,
+                            Description = "Fake description " + i
+                        };
+                        seedList.Add(product);
+
+
+
+                    }
+                    seedList.ForEach(n => ctx.TestProduct.Add(n));
+                    ctx.SaveChangesAsync();
+
+                }
             }
             catch(Exception e) {
                 Console.WriteLine(e.Message);
